+++ src/super/i-block/i-block.ts
@@ -150,7 +150,6 @@ export {
 	system,
 	watch,
 	wait,
-	bindModTo,
 	mod,
 	removeMod,
 	elMod,
@@ -1729,7 +1728,7 @@ export default abstract class iBlock extends ComponentInterface<iBlock, iStaticP
 	 * @param obj
 	 * @param [constructor] - component constructor
 	 */
-	protected isComponent<T extends iBlock>(obj: unknown, constructor?: {new(): T}): obj is T {
+	protected isComponent<T extends iBlock>(obj: unknown, constructor?: {new(): T} | Function): obj is T {
 		return Boolean(obj && (<Dictionary>obj).instance instanceof (constructor || iBlock));
 	}