+++ CHANGELOG.md
@@ -23,6 +23,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Renamed `data-block-params` -> `data-root-component-params` (`iStaticPage`)
 * Removed `bCheckboxGroup`
 * Removed `bCheckboxIcon`
+* Removed `@bindModTo` decorator
 
 ## v3.0.0-beta.207 (2019-11-07)